[SPARK-16584][SQL] Move regexp unit tests to RegexpExpressionsSuite
